rename folder to camelCase
