+++ src/components/main/Scrolls.tsx
@@ -5,7 +5,7 @@ import { ScrollMenu } from 'react-horizontal-scrolling-menu';
 import { LeftArrow, RightArrow } from './Arrow';
 import Block from './block/Block';
 import useRooms from '../../hooks/useRooms';
-import { filterData } from './filterData';
+import { filterBlock } from './block/filterBlock';
 import TagStore from '../../stores/searchInputStore';
 import EmptyBlock from './block/EmptyBlock';
 
@@ -13,7 +13,7 @@ export default function Scrolls() {
   const { isLoading, error, data } = useRooms();
   const { searchInput } = TagStore();
   if (error) return <div>An error has occurred: </div>;
-  const newData = filterData(data, searchInput);
+  const filteredBlock = filterBlock(data, searchInput);
 
   return (
     <Container>
@@ -22,7 +22,7 @@ export default function Scrolls() {
           ? [...Array(3)].map((no, index) => (
               <EmptyBlock key={Symbol(index).toString()} isMain={false} />
             ))
-          : newData.map((data) => {
+          : filteredBlock.map((data) => {
               return <Block key={data.itemId} isMain={false} data={data} />;
             })}
       </ScrollMenu>
