+++ src/components/profile/Contents.tsx
@@ -2,12 +2,12 @@ import { useState } from 'react';
 import styled from 'styled-components';
 import userStore from 'src/stores/userStore';
 import media from 'src/styles/media';
-import UserProfile from './UserProfile/UserProfile';
+import UserProfile from './userProfile/UserProfile';
 import Friends from './Friends';
 import Statistics from './Statistics';
 import Overall from './Overall';
 import roomModalStore from '../../stores/room/roomModalStore';
-import EditUserProfile from './EditUserProfile/EditUserProfile';
+import EditUserProfile from './editUserProfile/EditUserProfile';
 
 export default function Contents({
   userId,
