+++ src/components/profile/chart/calculateLabels.ts
