+++ src/components/profile/friends/AcceptOrDecline.tsx
