+++ src/components/profile/friends/FriendIcon.tsx
